stop_n must be int

+++ CIME/SystemTests/system_tests_common.py
@@ -173,7 +173,7 @@ class SystemTestsCommon(object):
         else:
             expect(False, f"stop_option {stop_option} not available for this test")
 
-        stop_n = stop_n * factor // coupling_secs
+        stop_n = int(stop_n * factor // coupling_secs)
         rest_n = math.ceil((stop_n // 2 + 1) * coupling_secs / factor)
 
         expect(stop_n > 0, "Bad STOP_N: {:d}".format(stop_n))
